Fix admin deploy audit Telegram ID overflow with BIGINT migration

+++ app/models/admin_deploy_audit.py
@@ -4,7 +4,7 @@ import uuid
 from datetime import datetime
 from typing import Optional, Any, Dict
 
-from sqlalchemy import DateTime, Integer, Text, JSON
+from sqlalchemy import BigInteger, DateTime, Text, JSON
 from sqlalchemy.dialects.postgresql import UUID
 from sqlalchemy.orm import Mapped, mapped_column
 
@@ -17,9 +17,9 @@ class AdminDeployAudit(Base):
     id: Mapped[uuid.UUID] = mapped_column(UUID(as_uuid=True), primary_key=True, default=uuid.uuid4)
     operation_id: Mapped[str] = mapped_column(Text, nullable=False, unique=True, index=True)
 
-    requested_by_tg_user_id: Mapped[Optional[int]] = mapped_column(Integer, nullable=True)
+    requested_by_tg_user_id: Mapped[Optional[int]] = mapped_column(BigInteger, nullable=True)
     requested_by_username: Mapped[Optional[str]] = mapped_column(Text, nullable=True)
-    chat_id: Mapped[int] = mapped_column(Integer, nullable=False)
+    chat_id: Mapped[int] = mapped_column(BigInteger, nullable=False)
 
     requested_at: Mapped[datetime] = mapped_column(DateTime(timezone=True), nullable=False)
     confirmed_at: Mapped[Optional[datetime]] = mapped_column(DateTime(timezone=True), nullable=True)
